chore(airplay_receiver): cleanup transport review findings
- Fix %d/int32_t format warning in volume log (cast to int)
- Dedupe USE_AIRPLAY_HEAP_TRACE: drop redundant add_build_flag, keep add_define
- Replace transport std::string device_name with a fixed buffer so ALL transport
  heap stays on the airplay_* path
- Remove dead code (pair-verify pvs local, rtsp_conn_set_volume empty event data,
  unused device_name_ member); clarify raw pair-verify fallback comment
- Refresh manifest description/version (transport ported, audio engine pending)

+++ components/airplay_receiver/airplay_receiver.cpp
@@ -58,7 +58,7 @@ void AirPlayReceiver::handle_transport_event(TransportEvent event, const Transpo
       }
       break;
     case TRANSPORT_EVENT_VOLUME:
-      ESP_LOGI(TAG, "transport: volume db=%.2f q15=%d", transport_volume_db(), transport_volume_q15());
+      ESP_LOGI(TAG, "transport: volume db=%.2f q15=%d", transport_volume_db(), (int) transport_volume_q15());
       break;
     case TRANSPORT_EVENT_PLAYING:
       ESP_LOGI(TAG, "transport: playing");
